Move RestIssueUpdateVersion::getJson() to RestBase
